feat: move functions to separate module
- move getNames and addName functions to server-functions.js
