+++ index.js
@@ -1,14 +1,6 @@
 import http from 'node:http';
-import fs from 'node:fs';
+import { getNames,addName } from './server-functions';
 
-async function getNames() {
-    return await fs.promises.readFile('names.txt', 'utf8');
-}
-
-async function addName(name) {
-    const names = await getNames();
-    await fs.promises.writeFile('names.txt', names + '\n' + name);
-}
 
 const server = http.createServer(async (req, res) => {
     if (req.url === '/names') {
